Commit with babel and hash
